rename sample app
