--- example-kt-01reactiveui/src/main/java/foo/bar/example/forereactiveui/App.kt
@@ -1,28 +1,0 @@
-package foo.bar.example.forereactiveui
-
-import android.app.Application
-import co.early.fore.kt.core.delegate.DebugDelegateDefault
-import co.early.fore.kt.core.delegate.ForeDelegateHolder
-
-/**
- * Copyright © 2015-2020 early.co. All rights reserved.
- */
-class App : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        if (BuildConfig.DEBUG){
-            ForeDelegateHolder.setDelegate(DebugDelegateDefault("fore_"))
-        }
-
-        inst = this
-
-        OG.setApplication(this)
-        OG.init()
-    }
-
-    companion object {
-        lateinit var inst: App private set
-    }
-}
