--- example01databinding/src/main/java/foo/bar/example/foredatabinding/App.java
@@ -1,26 +1,0 @@
-package foo.bar.example.foredatabinding;
-
-import android.app.Application;
-
-/**
- * Try not to fill this class with lots of code, if possible move it to a model somewhere
- */
-public class App extends Application {
-
-    private static App inst;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-
-        inst = this;
-
-        OG.setApplication(inst);
-        OG.init();
-    }
-
-    public static App getInst() {
-        return inst;
-    }
-
-}
