+++ example01reactiveui/src/main/AndroidManifest.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="foo.bar.example.forereactiveuikt">
+    package="foo.bar.example.forereactiveui">
 
     <application
-        android:name="foo.bar.example.forereactiveuikt.App"
+        android:name="foo.bar.example.forereactiveui.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="foo.bar.example.forereactiveuikt.ui.wallet.WalletsActivity">
+        <activity android:name="foo.bar.example.forereactiveui.ui.wallet.WalletsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -17,5 +17,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
