+++ example01reactiveui/src/main/res/layout/activity_wallet.xml
