+++ example01reactiveui/src/main/res/mipmap-hdpi/ic_launcher.png
