+++ example01reactiveui/src/main/res/mipmap-mdpi/ic_launcher.png
