+++ example01reactiveui/src/main/res/mipmap-mdpi/ic_launcher_background.png
