+++ example01reactiveui/src/main/res/mipmap-xhdpi/ic_launcher_foreground.png
