+++ example01reactiveui/src/main/res/mipmap-xxhdpi/ic_launcher_foreground.png
