+++ example01reactiveui/src/main/res/mipmap-xxxhdpi/ic_launcher_background.png
