+++ example01reactiveui/src/main/res/values/colors.xml
