+++ example01reactiveui/src/main/res/values/strings.xml
