+++ settings.gradle.kts
@@ -9,7 +9,7 @@ include(":fore-jv")
 include(":fore-kt")
 
 include(":example-kt-01reactiveui", ":example-kt-02coroutine", "example-kt-03adapters",
-        ":example-kt-04retrofit", ":example01databinding", ":example02threading", ":example03adapters",
+        ":example-kt-04retrofit", ":example01reactiveui", ":example02threading", ":example03adapters",
         ":example04retrofit", ":example05ui", ":example06db")
 
 rootProject.name = "android-fore"
